updating internals for Mojolicious 2.51

+++ extlib/Mojo/Asset/File.pm
@@ -4,7 +4,7 @@ use Mojo::Base 'Mojo::Asset';
 use Carp 'croak';
 use Errno;
 use Fcntl;
-use File::Copy ();
+use File::Copy 'move';
 use File::Spec;
 use IO::File;
 use Mojo::Util 'md5_sum';
@@ -13,28 +13,27 @@ has [qw/cleanup path/];
 has handle => sub {
   my $self = shift;
 
-  # Already got a file without handle
+  # Open existing file
   my $handle = IO::File->new;
-  my $file   = $self->path;
-  if ($file && -f $file) {
-    $handle->open("< $file")
-      or croak qq/Can't open file "$file": $!/;
+  my $path   = $self->path;
+  if (defined $path && -f $path) {
+    $handle->open("< $path") or croak qq/Can't open file "$path": $!/;
     return $handle;
   }
 
-  # Open existing or temporary file
+  # Open new or temporary file
   my $base = File::Spec->catfile($self->tmpdir, 'mojo.tmp');
-  my $name = $file || $base;
+  my $name = $path // $base;
   my $fh;
   until (sysopen $fh, $name, O_CREAT | O_EXCL | O_RDWR) {
-    croak qq/Can't open file "$name": $!/ if $file || $! != $!{EEXIST};
+    croak qq/Can't open file "$name": $!/
+      if defined $path || $! != $!{EEXIST};
     $name = "$base." . md5_sum(time . $$ . rand 9999999);
   }
-  $file = $name;
-  $self->path($file);
+  $self->path($name);
 
   # Enable automatic cleanup
-  $self->cleanup(1);
+  $self->cleanup(1) unless defined $self->cleanup;
 
   # Open for read/write access
   $handle->fdopen(fileno($fh), "+>") or croak qq/Can't open file "$name": $!/;
@@ -49,23 +48,18 @@ has tmpdir => sub { $ENV{MOJO_TMPDIR} || File::Spec->tmpdir };
 #  claws!"
 sub DESTROY {
   my $self = shift;
-  my $path = $self->path;
-  if ($self->cleanup && -f $path) {
-    close $self->handle;
-    unlink $path;
-  }
+  return unless $self->cleanup && defined(my $path = $self->path);
+  close $self->handle;
+  unlink $path if -w $path;
 }
 
 sub add_chunk {
   my ($self, $chunk) = @_;
-
-  # Seek to end
-  $self->handle->sysseek(0, SEEK_END);
-
-  # Append to file
+  my $handle = $self->handle;
+  $handle->sysseek(0, SEEK_END);
   $chunk //= '';
-  $self->handle->syswrite($chunk, length $chunk);
-
+  croak "Can't write to asset: $!"
+    unless defined $handle->syswrite($chunk, length $chunk);
   return $self;
 }
 
@@ -73,25 +67,26 @@ sub contains {
   my ($self, $pattern) = @_;
 
   # Seek to start
-  $self->handle->sysseek($self->start_range, SEEK_SET);
+  my $handle = $self->handle;
+  $handle->sysseek($self->start_range, SEEK_SET);
+
+  # Calculate window
   my $end = $self->end_range // $self->size;
   my $window_size = length($pattern) * 2;
   $window_size = $end - $self->start_range
     if $window_size > $end - $self->start_range;
-
-  # Read
-  my $read         = $self->handle->sysread(my $window, $window_size);
+  my $read         = $handle->sysread(my $window, $window_size);
   my $offset       = $read;
-  my $pattern_size = length($pattern);
+  my $pattern_size = length $pattern;
+  my $range        = $self->end_range;
 
   # Moving window search
-  my $range = $self->end_range;
   while ($offset <= $end) {
     if (defined $range) {
       $pattern_size = $end + 1 - $offset;
       return -1 if $pattern_size <= 0;
     }
-    $read = $self->handle->sysread(my $buffer, $pattern_size);
+    $read = $handle->sysread(my $buffer, $pattern_size);
     $offset += $read;
     $window .= $buffer;
     my $pos = index $window, $pattern;
@@ -108,21 +103,19 @@ sub get_chunk {
 
   # Seek to start
   $start += $self->start_range;
-  $self->handle->sysseek($start, SEEK_SET);
-  my $end = $self->end_range;
-  my $buffer;
-
-  # Chunk size
-  my $size = $ENV{MOJO_CHUNK_SIZE} || 131072;
+  my $handle = $self->handle;
+  $handle->sysseek($start, SEEK_SET);
 
   # Range support
-  if (defined $end) {
+  my $buffer;
+  my $size = $ENV{MOJO_CHUNK_SIZE} || 131072;
+  if (defined(my $end = $self->end_range)) {
     my $chunk = $end + 1 - $start;
     return '' if $chunk <= 0;
     $chunk = $size if $chunk > $size;
-    $self->handle->sysread($buffer, $chunk);
+    $handle->sysread($buffer, $chunk);
   }
-  else { $self->handle->sysread($buffer, $size) }
+  else { $handle->sysread($buffer, $size) }
 
   return $buffer;
 }
@@ -130,40 +123,32 @@ sub get_chunk {
 sub is_file {1}
 
 sub move_to {
-  my ($self, $path) = @_;
+  my ($self, $to) = @_;
 
   # Windows requires that the handle is closed
   close $self->handle;
   delete $self->{handle};
 
-  # Move
-  my $src = $self->path;
-  File::Copy::move($src, $path)
-    or croak qq/Can't move file "$src" to "$path": $!/;
-  $self->path($path);
-
-  # Don't clean up a moved file
-  $self->cleanup(0);
+  # Move file and prevent clean up
+  my $from = $self->path;
+  move($from, $to) or croak qq/Can't move file "$from" to "$to": $!/;
+  $self->path($to)->cleanup(0);
 
   return $self;
 }
 
 sub size {
   my $self = shift;
-  return 0 unless my $file = $self->path;
+  return 0 unless defined(my $file = $self->path);
   return -s $file;
 }
 
 sub slurp {
-  my $self = shift;
-
-  # Seek to start
-  $self->handle->sysseek(0, SEEK_SET);
-
-  # Slurp
+  my $self   = shift;
+  my $handle = $self->handle;
+  $handle->sysseek(0, SEEK_SET);
   my $content = '';
-  while ($self->handle->sysread(my $buffer, 131072)) { $content .= $buffer }
-
+  while ($handle->sysread(my $buffer, 131072)) { $content .= $buffer }
   return $content;
 }
 
@@ -178,11 +163,15 @@ Mojo::Asset::File - File storage for HTTP 1.1 content
 
   use Mojo::Asset::File;
 
+  # Temporary file
   my $file = Mojo::Asset::File->new;
   $file->add_chunk('foo bar baz');
+  say 'File contains "bar"' if $file->contains('bar') >= 0;
   say $file->slurp;
 
+  # Existing file
   my $file = Mojo::Asset::File->new(path => '/foo/bar/baz.txt');
+  $file->move_to('/yada.txt');
   say $file->slurp;
 
 =head1 DESCRIPTION
@@ -206,22 +195,23 @@ Delete file automatically once it's not used anymore.
   my $handle = $file->handle;
   $file      = $file->handle(IO::File->new);
 
-Actual file handle.
+File handle, created on demand.
 
 =head2 C<path>
 
   my $path = $file->path;
   $file    = $file->path('/foo/bar/baz.txt');
 
-Actual file path.
+File path used to create C<handle>, can also be automatically generated if
+necessary.
 
 =head2 C<tmpdir>
 
   my $tmpdir = $file->tmpdir;
   $file      = $file->tmpdir('/tmp');
 
-Temporary directory, defaults to the value of C<MOJO_TMPDIR> or auto
-detection.
+Temporary directory used to generate C<path>, defaults to the value of
+C<MOJO_TMPDIR> or auto detection.
 
 =head1 METHODS
